Replace SelectBundlePanel with BundleSelector witch uses CommonPagedPanel class.
